requirements.txt
amqp==2.5.2
certifi==2020.4.5.1
cffi==1.14.0
cryptography==2.9.2
Django==3.1.5
django-filter==1.1.0
django-silk==4.1.0
django-nested-admin==3.3.0
djangorestframework==3.12.2
#djangorestframework-filters==0.11.1
drf-api-tracking==1.7.8
gunicorn==20.0.4
idna==2.9
itsdangerous==1.1.0
loguru==0.5.0
mysqlclient==1.4.6
olefile==0.46
Pillow==7.1.2
pip==20.0.2
psycopg2-binary==2.8.5
pycparser==2.20
PyMySQL==0.9.3
python-monkey-business==1.0.0
pytz==2020.1
setuptools
simpleui==4.0.2
six==1.14.0
vine==1.3.0
wheel==0.34.2
gunicorn==20.0.4
